// src/app/model/product/product.component.ts
export class ProductComponents{
  position: number = 0;
  id: number = 0;
  nome: string = "";
  imagem: string = "";
  codigo_de_barras: string = "";
  preco: number = 0;
  quantidade: number = 0;
  fornecedor_id: number = 0;
  fornecedor_nome: string = "";
  categoria_id: number = 0;
  categoria_nome:string = "";
}